requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
numpy
opencv-python
Pillow
scikit-image
lpips
torchmetrics
tqdm
einops